Remove and refactor outdated query and delete book tests
Removed the obsolete test for querying books by ISBN and added a test to check that 'top' value of zero is invalid. Updated the 'top' field in the model to ensure it must be greater than or equal to one. Refactored the delete book test to use a more specific setup for book data.